+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 <!--
-SPDX-FileCopyrightText: 2023-2024 Antoine Belvire
+SPDX-FileCopyrightText: 2025 Antoine Belvire
 SPDX-License-Identifier: GPL-3.0-or-later
 -->
+
+## Not released
+
+- gui: Auto-advance current box when typing.
 
 ## Version 0.11 - Not much but still - 2024/12/30